ENH: Convert `barplot` to use DataFrame over Table objects
