+++ q2_taxa/_method.py
@@ -8,16 +8,7 @@
 
 import pandas as pd
 
-
-def _collapse_pd_table(table, taxonomy, level, max_observed_level):
-    def _collapse(tax):
-        tax = [x.strip() for x in tax.split(';')]
-        if len(tax) < max_observed_level:
-            padding = ['__'] * (max_observed_level - len(tax))
-            tax.extend(padding)
-        return ';'.join(tax[:level])
-    table.columns = taxonomy.apply(_collapse)[table.columns]
-    return table.groupby(table.columns, axis=1).agg(sum)
+from ._util import _collapse_table, _get_max_level
 
 
 def collapse(table: pd.DataFrame, taxonomy: pd.Series,
@@ -27,11 +18,11 @@ def collapse(table: pd.DataFrame, taxonomy: pd.Series,
                          'than or equal to 1.' % level)
 
     # Assemble the taxonomy data
-    max_observed_level = taxonomy.apply(lambda x: len(x.split(';'))).max()
+    max_observed_level = _get_max_level(taxonomy)
 
     if level > max_observed_level:
         raise ValueError('Requested level of %d is larger than the maximum '
                          'level available in taxonomy data (%d).' %
                          (level, max_observed_level))
 
-    return _collapse_pd_table(table, taxonomy, level, max_observed_level)
+    return _collapse_table(table, taxonomy, level, max_observed_level)
