+++ q2_taxa/_util.py
@@ -7,31 +7,27 @@
 # ----------------------------------------------------------------------------
 
 
-def _assemble_taxonomy_data(taxonomy, table):
-    taxa = {}
-    max_obs_lvl = 0
-    for k, v in taxonomy.iteritems():
-        levels = [x.strip() for x in v.split(';')]
-        max_obs_lvl = len(levels) if len(levels) > max_obs_lvl else max_obs_lvl
-        taxa[k] = {'taxonomy': levels}
-    table.add_metadata(taxa, axis='observation')
-
-    return table, max_obs_lvl
+def _get_max_level(taxonomy):
+    return taxonomy.apply(lambda x: len(x.split(';'))).max()
 
 
-def _collapse_table(table, level, max_observed_level):
-    def bin_f(id_, x):
-        if len(x['taxonomy']) < max_observed_level:
-            padding = ['__'] * (max_observed_level - len(x['taxonomy']))
-            x['taxonomy'].extend(padding)
-        return ';'.join(x['taxonomy'][:level])
-    return table.collapse(bin_f, norm=False, min_group_size=1,
-                          axis='observation')
+def _collapse_table(table, taxonomy, level, max_observed_level):
+    table = table.copy()
+
+    def _collapse(tax):
+        tax = [x.strip() for x in tax.split(';')]
+        if len(tax) < max_observed_level:
+            padding = ['__'] * (max_observed_level - len(tax))
+            tax.extend(padding)
+        return ';'.join(tax[:level])
+
+    table.columns = taxonomy.apply(_collapse)[table.columns]
+    return table.groupby(table.columns, axis=1).agg(sum)
 
 
 def _extract_to_level(taxonomy, table, max_level=None):
     # Assemble the taxonomy data
-    table, max_obs_lvl = _assemble_taxonomy_data(taxonomy, table)
+    max_obs_lvl = _get_max_level(taxonomy)
 
     if max_level is None:
         max_level = max_obs_lvl
@@ -44,7 +40,7 @@ def _extract_to_level(taxonomy, table, max_level=None):
     collapsed_tables = []
     # Collapse table at specified level
     for level in range(1, max_level + 1):
-        collapsed_table = _collapse_table(table, level, max_obs_lvl)
+        collapsed_table = _collapse_table(table, taxonomy, level, max_obs_lvl)
         collapsed_tables.append(collapsed_table)
 
     return collapsed_tables
